Fix problems with 3rd-party parsers.
Pass default settings to custom parser for included file.
When the "include" directive is used with a custom parser,
we must ensure the custom parser's default settings are present
in `document.settings`.

Fix error when determining the document metadata title from the
source path and the internal `source` attribute is ``None``.

git-svn-id: svn://svn.code.sf.net/p/docutils/code/trunk@10126 929543f6-e4f2-0310-98a6-ba3bd3dd1d04